+++ client/scripts/components/Admin/ListView/DisclosureTable.jsx
@@ -16,7 +16,7 @@
     along with this program.  If not, see <http://www.gnu.org/licenses/>
 */
 
-import React from 'react'; //eslint-disable-line no-unused-vars
+import React from 'react';
 import {merge} from '../../../merge';
 import {DisclosureTableRow} from './DisclosureTableRow';
 import {TableHeading} from './TableHeading';
@@ -83,6 +83,7 @@ export class DisclosureTable extends React.Component {
           type={disclosure.type}
           statusCd={disclosure.statusCd}
           submittedDate={disclosure.submitted_date}
+          revisedDate={disclosure.revised_date}
           searchTerm={this.props.searchTerm}
         />
       );