+++ client/scripts/components/Admin/ListView/DisclosureTableRow.jsx
@@ -62,6 +62,14 @@ export class DisclosureTableRow extends React.Component {
       }
     };
 
+    let dateToShow;
+    if (this.props.revisedDate !== null) {
+      dateToShow = `${formatDate(this.props.revisedDate)} (revised)`;
+    }
+    else {
+      dateToShow = formatDate(this.props.submittedDate);
+    }
+
     return (
       <div role="row" style={merge(styles.container, this.props.style)}>
         <span role="gridcell" style={merge(styles.value, styles.firstColumn)}>
@@ -76,7 +84,7 @@ export class DisclosureTableRow extends React.Component {
           {ConfigStore.getAdminDisclosureStatusString(this.props.statusCd)}
         </span>
         <span role="gridcell" style={merge(styles.value, styles.lastColumn)}>
-          {formatDate(this.props.submittedDate)}
+          {dateToShow}
         </span>
       </div>
     );