+++ CHANGELOG.md
@@ -5,6 +5,8 @@ V1.0.3
   saved. This will simply return them to that place in the process.
 * Fixed bug which prevented adding new relationships to existing entities
 * Fixed bugs with the next step button not enabling/disabling at the correct times
+* Now shows the revised date (and labels it as such) on the admin list view for
+  revised disclosures. Un-revised disclosures still show the submitted date.
 
 V1.0.2